Adding 'safe' versions of FluentAssertionResultExtensions ShouldBeFailure extension methods
